Move scala-xml related code to separate module
