+++ modules/core/shared/src/main/scala/eu/timepit/refined/string.scala
@@ -63,9 +63,6 @@ object string extends StringInference {
   /** Predicate that checks if a `String` is a parsable `BigDecimal`. */
   final case class ValidBigDecimal()
 
-  /** Predicate that checks if a `String` is well-formed XML. */
-  final case class Xml()
-
   /** Predicate that checks if a `String` is a valid XPath expression. */
   final case class XPath()
 
@@ -228,11 +225,6 @@ object string extends StringInference {
   object ValidBigDecimal {
     implicit def validBigDecimalValidate: Validate.Plain[String, ValidBigDecimal] =
       Validate.fromPartial(BigDecimal(_), "ValidBigDecimal", ValidBigDecimal())
-  }
-
-  object Xml {
-    implicit def xmlValidate: Validate.Plain[String, Xml] =
-      Validate.fromPartial(scala.xml.XML.loadString, "Xml", Xml())
   }
 
   object XPath {
